requirements.txt
Flask
Flask-Cors
psycopg2-binary
sqlalchemy
pyarrow
python-dotenv
google-auth
google-auth-oauthlib # Often needed by google.oauth2.id_token
requests # Often needed by google-auth libraries
gunicorn # Production WSGI server
PyJWT
cryptography
pandas
matplotlib
marimo
